Simplify API calls by implicitly passing the required interface instance
Most of the "interfaces" defined in Steamworks are entry points for entire APIs such as matchmaking, user info, friends list, etc.

Every time you wanted to use them, you had to construct an instance of the interface and pass it to each call.

Now it's created for you, and you have one less argument to worry about. This breaks all existing code, but oh well, we're prototyping!!!!!!